Remove the mock method for formatter test
